+++ R/get_emr.R
@@ -8,11 +8,7 @@
 #'
 #' @name get_emr
 #'
-#' @param codes_df A data frame. Contains the `vocab_col` and `codes_col` i.e., a list of diagnostic codes, and an indicator of the vocabulary.
-#' @param vocab_col A string. Column name in `codes_df` that contains the vocabulary indicator for the code (vocab IDs should be 'ICD10' 'Read2' or 'CTV3').
-#'        \code{default='vocab_id'}
-#' @param codes_col A string. Column name in `codes_df` that contains the diagnostic code ('ICD10' 'Read2' or 'CTV3' codes).
-#'        \code{default='code'}
+#' @param codes_df A data frame. Contains two columns: `code` and `vocab_id` i.e., a list of diagnostic codes, and an indicator of the vocabulary. Other columns are ignored.
 #' @param spark_master A string. The `master` argmuent passed to `sparklyr::spark_connect()`.
 #'        \code{default='spark://master:41000'}
 #' @param verbose Logical. Be verbose,
@@ -36,21 +32,22 @@
 #'
 get_emr <- function(
 	codes_df,
-	vocab_col = "vocab_id",
-	codes_col = "code",
 	spark_master = "spark://master:41000",
 	verbose=FALSE
 )  {
 	
 	start_time <- Sys.time()
 
+	vocab_col = "vocab_id"
+	codes_col = "code"
+
 	# Check input
 	if (verbose) cat("Check inputs\n")
 	if (! any(class(codes_df) %in% c("data.frame","tbl_df")))  stop("Codelist needs to be provided as a data frame")
 	codes_df = as.data.frame(codes_df)  # in case a tibble
 
-	if (! vocab_col %in% colnames(codes_df))  stop("Codelist data frame needs to include vocabulary column. Specify with `vocab_col`")
-	if (! codes_col %in% colnames(codes_df))  stop("Codelist data frame needs to include codes column. Specify with `codes_col`")
+	if (! vocab_col %in% colnames(codes_df))  stop("Codelist data frame needs to include vocabulary column `vocab_id`")
+	if (! codes_col %in% colnames(codes_df))  stop("Codelist data frame needs to include codes colum `code`")
 
 	if (! any(c("ICD10","Read2","CTV3") %in% codes_df[,vocab_col]))  stop("Vocabularies need to include ICD10, Read2 or CTV3")